Display keystrokes with command keys correctly
Unsure whether this is reproducible on other systems, but on mine, any time
I would press a key with a command key (shift, option, command), it would
disply as

option + ^A (key)

Which was pretty annoying. I poked around in the keystroke transformer but
the code there is mostly taking note of the command keys and inserting them
into the string to be displayed and not doing anything crazy with the
remainder, so I assume this how the key + command character is actually
coming across.